Split fixedReadTokenFactory to separate module, improve function definitions
I’ve changed some functions from arrows to regular named functions.
Autobinding was not required, and named functions are easier to debug
when needed.
