+++ src/HtmlParser.js
@@ -1,5 +1,6 @@
 import * as supports from './supports';
 import * as streamReaders from './streamReaders';
+import fixedReadTokenFactory from './fixedReadTokenFactory';
 
 // Order of detection matters: detection of one can only
 // succeed if detection of previous didn't
@@ -10,199 +11,6 @@ const detect = {
   startTag: /^</,
   chars: /^[^<]/
 };
-
-// Detection has already happened when a reader is called.
-const reader = {
-  comment: stream => {
-    const index = stream.indexOf('-->');
-    if (index >= 0) {
-      return {
-        content: stream.substr(4, index - 1),
-        length: index + 3
-      };
-    }
-  },
-
-  endTag: stream => {
-    const match = stream.match(endTag);
-
-    if (match) {
-      return {
-        tagName: match[1],
-        length: match[0].length
-      };
-    }
-  },
-
-  atomicTag: stream => {
-    const start = reader.startTag(stream);
-    if (start) {
-      const rest = stream.slice(start.length);
-      // for optimization, we check first just for the end tag
-      if (rest.match(new RegExp('<\/\\s*' + start.tagName + '\\s*>', 'i'))) {
-        // capturing the content is inefficient, so we do it inside the if
-        const match = rest.match(new RegExp('([\\s\\S]*?)<\/\\s*' + start.tagName + '\\s*>', 'i'));
-        if (match) {
-          // good to go
-          return {
-            tagName: start.tagName,
-            attrs: start.attrs,
-            content: match[1],
-            length: match[0].length + start.length
-          };
-        }
-      }
-    }
-  },
-
-  startTag: stream => {
-    const endTagIndex = stream.indexOf('>');
-    if (endTagIndex === -1) {
-      return null; // avoid the match statement if there will be no match
-    }
-
-    const match = stream.match(startTag);
-    if (match) {
-      const attrs = {};
-      const booleanAttrs = {};
-      let rest = match[2];
-
-      match[2].replace(attr, function(match, name) {
-        if (!(arguments[2] || arguments[3] || arguments[4] || arguments[5])) {
-          attrs[name] = '';
-        } else if (arguments[5]) {
-          attrs[arguments[5]] = '';
-          booleanAttrs[arguments[5]] = true;
-        } else {
-          attrs[name] = arguments[2] || arguments[3] || arguments[4] ||
-            fillAttr.test(name) && name || '';
-        }
-
-        rest = rest.replace(match, '');
-      });
-
-      return {
-        tagName: match[1],
-        attrs,
-        booleanAttrs,
-        rest: rest.replace(/^[\s\uFEFF\xA0]+|[\s\uFEFF\xA0]+$/g, ''),
-        unary: !!match[3],
-        length: match[0].length
-      };
-    }
-  },
-
-  chars: stream => {
-    const index = stream.indexOf('<');
-    return {
-      length: index >= 0 ? index : stream.length
-    };
-  }
-};
-
-function fixedReadTokenFactory(parser, options, readTokenImpl) {
-  // Empty Elements - HTML 4.01
-  const EMPTY = /^(AREA|BASE|BASEFONT|BR|COL|FRAME|HR|IMG|INPUT|ISINDEX|LINK|META|PARAM|EMBED)$/i;
-
-  // Elements that you can| intentionally| leave open
-  // (and which close themselves)
-  const CLOSESELF = /^(COLGROUP|DD|DT|LI|OPTIONS|P|TD|TFOOT|TH|THEAD|TR)$/i;
-
-  const stack = [];
-  stack.last = function() {
-    return this[this.length - 1];
-  };
-
-  stack.lastTagNameEq = function(tagName) {
-    const last = this.last();
-    return last && last.tagName &&
-      last.tagName.toUpperCase() === tagName.toUpperCase();
-  };
-
-  stack.containsTagName = function(tagName) {
-    for (let i = 0, tok; (tok = this[i]); i++) {
-      if (tok.tagName === tagName) {
-        return true;
-      }
-    }
-    return false;
-  };
-
-  function correct(tok) {
-    if (tok && tok.type === 'startTag') {
-      tok.unary = EMPTY.test(tok.tagName) || tok.unary;
-      tok.html5Unary = !(/\/>$/).test(tok.text);
-    }
-    return tok;
-  }
-
-  function peekToken() {
-    const tmp = parser.stream;
-    const tok = correct(readTokenImpl());
-    parser.stream = tmp;
-    return tok;
-  }
-
-  function closeLast() {
-    const tok = stack.pop();
-
-    // prepend close tag to stream.
-    parser.prepend(`</${tok.tagName}>`);
-  }
-
-  const handlers = {
-    startTag: tok => {
-      const tagName = tok.tagName;
-
-      if (tagName.toUpperCase() === 'TR' && stack.lastTagNameEq('TABLE')) {
-        parser.prepend('<TBODY>');
-        prepareNextToken();
-      } else if (options.selfCloseFix && CLOSESELF.test(tagName) &&
-          stack.containsTagName(tagName)) {
-        if (stack.lastTagNameEq(tagName)) {
-          closeLast();
-        } else {
-          parser.prepend(`</${tok.tagName}>`);
-          prepareNextToken();
-        }
-      } else if (!tok.unary) {
-        stack.push(tok);
-      }
-    },
-
-    endTag: tok => {
-      const last = stack.last();
-      if (last) {
-        if (options.tagSoupFix && !stack.lastTagNameEq(tok.tagName)) {
-          // cleanup tag soup
-          closeLast();
-        } else {
-          stack.pop();
-        }
-      } else if (options.tagSoupFix) {
-        // cleanup tag soup part 2: skip this token
-        skipToken();
-      }
-    }
-  };
-
-  function skipToken() {
-    readTokenImpl();
-    prepareNextToken();
-  }
-
-  function prepareNextToken() {
-    const tok = peekToken();
-    if (tok && handlers[tok.type]) {
-      handlers[tok.type](tok);
-    }
-  }
-
-  return () => {
-    prepareNextToken();
-    return correct(readTokenImpl());
-  };
-}
 
 export default class HtmlParser {
   constructor(stream = '', options = {}) {
